moving source to source schema and creating gov_official_trades view

+++ helpers/utils.py
@@ -12,8 +12,8 @@ def fetch_with_retries(url, max_retries=10, initial_delay=12):
             # Retry with exponential backoff
             # Add jitter for API calls to be randomly staggered
             jittered_delay = random.uniform(initial_delay-1, initial_delay)
-            expo = 2 ** retries
-            wait_time = int(response.headers.get("Retry-After", jittered_delay * expo))
+            expo = int(2 ** retries)
+            wait_time = float(response.headers.get("Retry-After", jittered_delay * expo))
             print(f"Rate Limited: Retrying in {wait_time:.2f} seconds... Function Retries Left: {max_retries - retries}")
             time.sleep(wait_time)
             retries += 1
